Work around a bug that resets &cpoptions
If the filetype is determined by scripts.vim, it resets &cpo to the
default value, sets the filetype, then reset &cpo to it's previous
value. That will undo any changes made to &cpo from the ftplugin.

The recent migration to Vim9 script has unintentionally fixed this bug
for future releases.